rework: refactor luv_handle_t
introduces a new set of internal utilities which take
advantage of noreturn and compiler warning for unreachable
code.

merges lhandle.c and handle.c into a single handle.c

all handles are now stored directly in the userdata linearly
in the form `{ luv_handle_t, uv_handle_t, extra_data }`.
this allows us to avoid any pointer chasing and places all
handle information in a single contiguous block of memory.

`luv_newuserdata()` was moved to internal.c and now creates
a userdata directly with the given size instead of creating
a userdata to a pointer to our handle.

`luv_checkudata()` was removed in favor of
`luaL_checkudata()`

`luv_error()` is a replacement for `luaL_error()` that marks
the code path as unreachable. This is a macro that behaves
as if it were a noreturn function.

`luv_argerror()` is a replacement for `luaL_argerror()` with
extra functionality for formatting a message. This function
is marked as noreturn.

`luv_typeerror()` is a replacement for `luaL_typeerror()`
and `luv_arg_type_error()` that throws an error about the
type of an argument being incorrect. This function is marked
as noreturn.

`luv_checkopttype()` behaves like `luaL_checktype()` but
returns false if the value is nil instead of erroring.

`luv_error()` renamed to `luv_pushfail()` to better match
the naming convention of the other functions and how they
interact with Lua.

`luv_result()` renamed to `luv_pushresult()` and now accepts
an argument to determine whether a successful result should
be pushed as the status integer or as a boolean true.

`luv_status()` renamed to `luv_pushstatus()` to better match
the naming convention of the other functions and how they
interact with Lua.

`luv_is_callable()` renamed to `luv_iscallable()` to better
match the associated Lua functions.

`luv_check_callable()` renamed to `luv_checkcallable()` to
better match the associated Lua functions.

callbacks are now referenced to as either `closed` or
`event` instead of specializing `event` for every handle
type since each handle only has only event callback. this is
enumerated to prevent any confusion.

`luv_handle_of()` returns the specialized `uv_handle_t*`
(like `uv_stream_t*`) associated with the given
`luv_handle_t*`. This is necessary to go from userdata ->
uv_handle_t when coming from Lua.

`luv_handle_from()` returns the `luv_handle_t*` associated
(via subtracting the offset of the handle) with the given
`uv_handle_t*`. This is necessary to go from uv_handle_t ->
userdata when coming from a libuv callback.

`luv_handle_extra()` returns the `void*` associated with the
specialized `uv_handle_t*` given its `luv_handle_t*`.

`luv_handle_metatable()` was introduced to streamline the
metatable creation process for handles and ensure that
`luv_check_handle()` can identify any userdata as a handle
or not correctly.

`luv_setup_handle()` has become `luv_handle_create()` and is
now responsible for creating the userdata and preparing the
handle for use.

`luv_handle_unref()` is mostly unchanged, but now it also
sets the callback references to `LUA_NOREF` just in case.

`luv_find_handle()` was renamed `luv_handle_push()` to
better reflect its purpose.

`luv_handle_check_valid()` was introduced to check if a
handle is still valid and has not been closed because libuv
functions may only be called on unclosed handles.

`luv_check_callback()` was renamed to
luv_handle_set_callback() to better reflect its purpose.

`luv_handle_unset_callback()` was introduced to undo the
references set by luv_handle_set_callback(), particularly
when the function that would call the callback returned an
error.

`luv_call_callback()` was renamed to
luv_handle_send_callback() to better reflect its
relationship with the rest of the luv_handle_t interface.

`luv_check_handle()` now returns the luv_handle_t and checks
if the given userdata is a handle by checking a raw pointer
field instead of a global table lookup.

`luv_handle_tostring()` now indicates whether or not the
handle has been closed or not with a `(closed)` suffix.

`luv_handle_gc()` now resurrects the userdata if the handle
needs to be closed to allow the close callback to arrive.
This was not previously necessary because the handle's
lifetime was not bound to the userdata.

`luv_close()`, `luv_is_active()`, `luv_is_closing()`,
`luv_ref()`, `luv_unref()`, `luv_has_ref()`,
`luv_send_buffer_size()`, `luv_recv_buffer_size()` and
`luv_fileno()` now throw an error if called on a closed
handle.